きちゃないねぇ
コードをまとめてみました！（オブジェクト指向）
