--- numeron project/Assets/scripts/NumberChoice.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fa6ea20cbb1ad9a40898d941a2c14924
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
