--- numeron project/Assets/scripts/NumberChoice2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 88777e4e674a63f49be2251d933312dd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
