--- numeron project/Assets/scripts/NumberChoice3.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 71f1e5ad4ee38da4b852ab3f30bb76c3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
